feat: rename project:deploy:functions to deploy:functions
